Creation of Login and registration forms
Enter first name, last name, email and password. They have parameters that must be met in order to register. If you are already registered, you can apply immediately. Of course, the email and password are checked to see if you have already registered before so that you can log in.