Create a Login form and new Screen
I have improved the previous Login form, I have added additional checks, places where the user is created and places where it is checked whether the user is checking in the system. If it is successful, a new window is opened for the user, if it is not, the first window shows messages that it is not good.